+++ src/sender/mod.rs
@@ -3,3 +3,6 @@ pub use sender::*;
 
 pub mod smtp;
 pub use smtp::*;
+
+pub mod sendmail;
+pub use sendmail::*;